Updated for deployer updates
In the release process of SCDF, it was identified that Spring Cloud
Deployer had been upgraded to JDK 1.8 which was not desirable at the
given time.  As part of the rollback, the deployer implementations
provided remained at 1.8, but the SPI was rolled back to 1.7.  This
commit addresses the difference between the different versions as well
as updates the code to use the latest `TaskLauncher` code.