Updated the migration scripts to use all caps for table names (#288)
* Updated the migration scripts to use all caps for table names

In the original migration scripts the table names were all lower case.  However the original table definitions were capitalized, and on file systems that are case sensitive,
mysql fails to update the table.  Because of the 2 different cases.

resolves #287

* Updated the 1.1.x branch migration scripts

In the original migration scripts the table names were all lower case.  However the original table definitions were capitalized, and on file systems that are case sensitive,
mysql fails to update the table.  Because of the 2 different cases.

* Added migration script for db2.