Fixed build after 2.0 PR merge
The 2.0 PR merge missed a couple items that broke the build.
Specifically the maven plugin repository was not specified so the
snapshots for the Spring Boot Maven Plugin could not be found.  Also the
ordering of the parameters in the Hibernate generated SQL was different
so the JpaApplicationTests were failing.