Fixes documentation publication
This commit fixes the generation of the zip file used to autopublish the
documentation.  The previous version only included the HTML, epub, and
PDF files for the reference documentation.  However it was missing the
images and css for the HTML as well as all of the javadocs.  This commit
now fixes that so the javadoc as well as all the resources required for
the reference documentation is included in the zip file.

Resolves #327