TaskJobLauncherCommandLineRunner now inherits from JobLauncherCommandLineRunner
Used to be a copy of.  No just inherits.

Added feature check for completed/failed jobs before throwing exception

Currently in Task if the user sets failOnJobFailure and they are using a TaskExecutor it will not throw an exception when task fails.
This because at evaluation time the job has not started.
With the latest change Task waits for all jobs to complete before deciding whether an exception needs to be thrown or not.