Removed failOnJobFailureWaitTime feature.
This feature would kill a task that had batch jobs that ran longer than the configured time.
The issue is that no where in the batch ecosystem to we kill batch jobs forceably with exceptions.  We suggest a stop, but that goes beyond the scope of what this feature was to provide.  And its up to the user to soft land those long running tasks, by stopping them.